+++ R/01-plotFormatting.R
@@ -1,32 +1,36 @@
 getStyleForIndividuals <- function(pointStyleList, input) {
   # pointStyleList are reactive values -> lapply over the names and not(!) the list itself
-  style <- lapply(names(pointStyleList), function(x) {pointStyleList[[x]][input]}) %>% 
+  style <- lapply(names(pointStyleList), function(individual) {pointStyleList[[individual]][input]}) %>% 
     unlist()
   names(style) <- names(pointStyleList)
   
   style
 }
 
-getDefaultPointFormatForModels <- function(pointStyleList, modelNames) {
+# Get the default point format for models
+# 
+# @param pointStyleList list of point styles for each model
+# @param modelNames names of the models
+# @param isFullReset if TRUE, all point styles are reset to the default values
+getDefaultPointFormatForModels <- function(pointStyleList, modelNames, isFullReset = FALSE) {
   # default colours
   defaultColours <- ggplot2::scale_color_discrete()$palette(n = length(modelNames))
   names(defaultColours) <- modelNames
   
   # setup lists with default values for style specs
   for (i in modelNames) {
-    if (is.null(pointStyleList[[i]])) 
+    if (is.null(pointStyleList[[i]]) || isFullReset) {
       pointStyleList[[i]] <- config()[["defaultPointStyle"]][["dataPoints"]]
-    # use default colour per model
-    pointStyleList[[i]]["color"] <- defaultColours[i]
+      # use default colour per model
+      pointStyleList[[i]]["color"] <- defaultColours[i]
+    }
   }
   
   return(pointStyleList)
 }
 
 getDefaultTextFormat <- function() {
-  list(legendTitle  = config()[["defaultIntervalTimePlotTitle"]],
-       legendText  = config()[["defaultIntervalTimePlotText"]],
-       plotTitle  = config()[["defaultIntervalTimePlotTitle"]],
+  list(plotTitle  = config()[["defaultIntervalTimePlotTitle"]],
        xAxisTitle = config()[["defaultIntervalTimePlotTitle"]],
        yAxisTitle = config()[["defaultIntervalTimePlotTitle"]],
        yAxisTitle2 = config()[["defaultIntervalTimePlotTitle"]],