+++ R/NAMESPACE.R
@@ -25,9 +25,9 @@
 #' @importFrom rstan sampling extract
 #' @importFrom shinycssloaders withSpinner
 #' @importFrom shinyjs alert enable runjs
-#' @importFrom shinyTools addCustomPointsToGGplot calculateRescalingFactors customPointsServer 
-#'  customPointsUI dataExportButton dataExportServer extractTitle
-#'  formatPointsOfGGplot formatScalesOfGGplot formatTitlesOfGGplot
+#' @importFrom shinyTools addCustomPointsToGGplot applyLayoutServer applyLayoutUI
+#'  calculateRescalingFactors customPointsServer customPointsUI dataExportButton dataExportServer 
+#'  extractTitle formatPointsOfGGplot formatScalesOfGGplot formatTitlesOfGGplot
 #'  plotExportButton plotExportServer plotLegendServer plotLegendUI plotPointsServer plotPointsUI
 #'  plotRangesServer plotRangesUI plotTitlesServer plotTitlesUI setLegendThemeOfGGplot
 #'  shinyTryCatch