+++ NEWS.md
@@ -1,5 +1,12 @@
 # OsteoBioR 24.11.0
 
+## New Features
+- option to add custom points to a plot
+- option to use custom legend labels for the plot
+- new logos and links in header
+
+# OsteoBioR 24.11.0
+
 ## Updates
 - integration of modules from _ChangeR_ package for break-point detection (#72)
 - option to set the "x" and "y" columns for mcp modelling